Factor out NWP preprocessing functions and add regridding
