+++ pyproject.toml
@@ -41,6 +41,7 @@ dependencies = [
     "grpclib>=0.4.7",
     "dp-sdk",
     "gcsfs==2025.10.0",
+    "xesmf==0.9.2", # Requires ESMF to be installed
 ]
 
 [dependency-groups]
