+++ site_forecast_app/models/pvnet/utils.py
@@ -6,11 +6,9 @@ import os
 import fsspec
 import numpy as np
 import torch
-import xarray as xr
 import yaml
-from ocf_data_sampler.config.model import NWP, Configuration
+from ocf_data_sampler.config.model import Configuration
 from ocf_data_sampler.config.save import save_yaml_configuration
-from pydantic import BaseModel
 
 from .consts import (
     generation_path,
@@ -21,15 +19,6 @@ from .consts import (
 )
 
 log = logging.getLogger(__name__)
-
-
-class NWPProcessAndCacheConfig(BaseModel):
-    """Configuration for processing and caching NWP data."""
-
-    source_nwp_path: str
-    dest_nwp_path: str
-    source: str
-    config: NWP | None = None
 
 
 def populate_data_config_sources(input_path: str, output_path: str) -> dict:
@@ -130,65 +119,6 @@ def populate_data_config_sources(input_path: str, output_path: str) -> dict:
     save_yaml_configuration(configuration, output_path)
 
     return config
-
-
-def process_and_cache_nwp(nwp_config: NWPProcessAndCacheConfig) -> None:
-    """Reads zarr file, renames t variable to t2m and saves zarr to new destination."""
-    source_nwp_path = nwp_config.source_nwp_path
-    dest_nwp_path = nwp_config.dest_nwp_path
-
-    log.info(
-        f"Processing and caching NWP data for {source_nwp_path} "
-        f"and saving to {dest_nwp_path} for {nwp_config.source}",
-    )
-
-    if os.path.exists(dest_nwp_path):
-        log.info(f"File already exists at {dest_nwp_path}")
-        return
-
-    # Load dataset from source
-    ds = xr.open_zarr(source_nwp_path, consolidated=False).load()
-
-    varname = next(iter(ds.data_vars))
-    if np.isnan(ds[varname].values).any():
-        raise ValueError(f"Found NaNs in {source_nwp_path}")
-
-    # This is important to avoid saving errors
-    for v in list(ds.coords.keys()):
-        ds[v].encoding.clear()
-
-    for v in list(ds.variables.keys()):
-        ds[v].encoding.clear()
-
-    if "variable" in ds.coords:
-        # make the dtype of variables is strings
-        ds["variable"] = ds.variable.astype(str)
-
-    if "channel" in ds.coords:
-        # make the dtype of variables is strings
-        ds["channel"] = ds.channel.astype(str)
-
-
-    # lets make sure its in the right order
-    if "init_time" in ds.coords:
-        ds = ds.transpose("init_time", "step", "variable", "latitude", "longitude")
-
-    scale_mo_global_clouds = os.getenv("MO_GLOBAL_SCALE_CLOUDS", "1") == "1"
-    if nwp_config.source == "mo_global" and scale_mo_global_clouds:
-        log.warning("Scaling MO Global cloud variables by from 0-100 to 0-1")
-
-        cloud_vars = [
-            "cloud_cover_high",
-            "cloud_cover_low",
-            "cloud_cover_medium",
-        ]
-        for cloud_var in cloud_vars:
-            idx = list(ds.variable.values).index(cloud_var)
-            ds[varname][:, :, idx] = ds[varname][:, :, idx] / 100.0
-
-    # Save destination path
-    log.info(f"Saving NWP data to {dest_nwp_path}")
-    ds.to_zarr(dest_nwp_path, mode="a")
 
 
 def set_night_time_zeros(
